arroz

+++ Site_SmartComfort/Models/Produto.cs
@@ -1,9 +1,9 @@
-﻿namespace Site_SmartComfort.Models
+﻿    namespace Site_SmartComfort.Models
 {
     public class Produto
     {
         public int Id { get; set; }
-        public decimal CodBar { get; set; }
+        public long CodBar { get; set; }
         public string NomePro { get; set; }
         public decimal PrecoPro { get; set; }
         public int QtdEstoquePro { get; set; }
